[feat] 코드 리팩토링 : 유저가디언과 유저와드의 중복된 코드를 한곳에 묶도록 변경
